financialia/requirements.txt
fastapi
uvicorn
keras
joblib
pydantic
asyncpg
databases
psycopg2-binary
tensorflow
numpy
pandas
scikit-learn
matplotlib
yfinance
jinja2
requests
nltk
python-multipart
python-dotenv